Move TypeConverter to core-base/database module
