+++ core-base/database/src/commonMain/kotlin/org/mifos/corebase/TypeConverter.kt
@@ -7,8 +7,9 @@
  *
  * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
  */
-package org.mifos.core.database
+package org.mifos.corebase
 
-import androidx.room.TypeConverter
-
-actual typealias TypeConverter = TypeConverter
+@Suppress("NO_ACTUAL_FOR_EXPECT")
+@Target(AnnotationTarget.FUNCTION)
+@Retention(AnnotationRetention.BINARY)
+expect annotation class TypeConverter()
