+++ core-base/database/src/desktopMain/kotlin/org/mifos/corebase/TypeConverter.desktop.kt
@@ -7,7 +7,7 @@
  *
  * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
  */
-package org.mifos.core.database
+package org.mifos.corebase
 
 import androidx.room.TypeConverter
